Make sample.frames always non-null.
In filtered samples, null frames (sample.frames == null) and
null samples (sample == null) seem to be treated the same,
except for the cases where the null frames are not handled at all.

This changes chargeNonJSToCallers to null out the sample instead
of frames, so that sample.frames is always non-null.

This fixes an issue with missing sample.frames null-check
in filterByCallstackPrefix() by making it non-necessary.
The missing null-check caused an exception when turning on the
"Javascript only" filter while focusing (via
FocusedCallstackPrefixSampleFilter)
on a portion of a stack that didn't have any JS in it.

I couldn't find explanation of why the null frames were
needed in the first place in the commits that initially
added this code:
https://github.com/bgirard/cleopatra/commit/055a8dc3d841d22cf0cb46bc37e242a9379ad776
https://github.com/bgirard/cleopatra/commit/7162f031adcbcb88b20212a8102ba86c39182752